pyproject.toml
[tool.poetry]
name = "shurutech"
version = "0.1.0"
description = ""
authors = ["Pratik Pathak <pratikpathak200@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
